+++ pykilosort/main.py
@@ -1,6 +1,5 @@
 import logging
 import shutil
-import os
 from pathlib import Path, PurePath
 
 import numpy as np
@@ -8,7 +7,7 @@ import numpy as np
 from .preprocess import preprocess, get_good_channels, get_whitening_matrix, get_Nbatch, destriping
 from .cluster import clusterSingleBatches
 from .datashift2 import datashift2
-from .learn import learnAndSolve8b, compress_templates
+from .learn import learnAndSolve8b
 from .postprocess import find_merges, splitAllClusters, set_cutoff, rezToPhy
 from .utils import Bunch, Context, memmap_large_array, load_probe, copy_bunch, DataLoader, \
                 RawDataLoader
@@ -33,12 +32,8 @@ def run(
 ):
     """Launch KiloSort 2.
 
-    probe has the following attributes:
-    - xc
-    - yc
-    - kcoords
-    - Nchan
-
+    stop_after: str
+        Stop after the given step. Possible values are: whitening_matrix, preprocess, drift_correction, learn, merge, split_1, cutoff
     """
 
     # Get or create the probe object.
@@ -94,29 +89,11 @@ def run(
     # Find good channels.
     # NOTE: now we use C order from loading up to the creation of the proc file, which is
     # in Fortran order.
-    params.minfr_goodchannels = 0
-    if params.minfr_goodchannels > 0:  # discard channels that have very few spikes
-        if "good_channels" not in ctx.timer.keys():
-            # determine bad channels
-            with ctx.time("good_channels"):
-                ir.igood = get_good_channels(
-                    raw_data=raw_data, probe=probe, params=params
-                )
-            # Cache the result.
-            ctx.write(igood=ir.igood)
-        if stop_after == "good_channels":
-            return ctx
-
-        # it's enough to remove bad channels from the channel map, which treats them
-        # as if they are dead
-        ir.igood = ir.igood.ravel().astype("bool")
-        probe.chanMap = probe.chanMap[ir.igood]
-        probe.xc = probe.xc[ir.igood]  # removes coordinates of bad channels
-        probe.yc = probe.yc[ir.igood]
-        probe.kcoords = probe.kcoords[ir.igood]
-    probe.Nchan = len(
-        probe.chanMap
-    )  # total number of good channels that we will spike sort
+    probe.Nchan = len(probe.chanMap)
+    # keep all channels but still output a quality control metric. Also the covariance matrix will be stabilized
+    # by taking into account those noisy channels
+    probe.good_channels, probe.channels_labels = get_good_channels(
+        raw_data, params, probe, method='raw_correlations', return_labels=True)
     assert probe.Nchan > 0
 
     # upper bound on the number of templates we can have
@@ -129,7 +106,7 @@ def run(
         # of the data
         with ctx.time("whitening_matrix"):
             ir.Wrot = get_whitening_matrix(
-                raw_data=raw_data, probe=probe, params=params
+                raw_data=raw_data, probe=probe, params=params, qc_path=ctx_path
             )
         # Cache the result.
         ctx.write(Wrot=ir.Wrot)