+++ pykilosort/preprocess.py
@@ -1,14 +1,16 @@
 import logging
 from math import ceil
-from functools import lru_cache
+import matplotlib.pyplot as plt
 
 import numpy as np
 from scipy.signal import butter
+import scipy.stats
 import cupy as cp
 from tqdm.auto import tqdm
 
+import pykilosort.qc
 from .cptools import lfilter, median
-from neurodsp.voltage import decompress_destripe_cbin
+from neurodsp.voltage import decompress_destripe_cbin, destripe, detect_bad_channels
 
 logger = logging.getLogger(__name__)
 
@@ -135,7 +137,7 @@ def my_sum(S1, sig, varargin=None):
     return S1
 
 
-def whiteningFromCovariance(CC):
+def whiteningFromCovariance(CC, epsilon=1e-6):
     # function Wrot = whiteningFromCovariance(CC)
     # takes as input the matrix CC of channel pairwise correlations
     # outputs a symmetric rotation matrix (also Nchan by Nchan) that rotates
@@ -143,19 +145,17 @@ def whiteningFromCovariance(CC):
 
     # covariance eigendecomposition (same as svd for positive-definite matrix)
     E, D, _ = cp.linalg.svd(CC)
-    eps = 1e-6
-    Di = cp.diag(1. / (D + eps) ** .5)
+    Di = cp.diag(1. / (D + epsilon) ** .5)
     Wrot = cp.dot(cp.dot(E, Di), E.T)  # this is the symmetric whitening matrix (ZCA transform)
     return Wrot
 
 
-def whiteningLocal(CC, yc, xc, nRange):
+def whiteningLocal(CC, yc, xc, nRange, epsilon=1e-6):
     # function to perform local whitening of channels
     # CC is a matrix of Nchan by Nchan correlations
     # yc and xc are vector of Y and X positions of each channel
     # nRange is the number of nearest channels to consider
     Wrot = cp.zeros((CC.shape[0], CC.shape[0]))
-
     for j in range(CC.shape[0]):
         ds = (xc - xc[j]) ** 2 + (yc - yc[j]) ** 2
         ilocal = np.argsort(ds)
@@ -163,85 +163,152 @@ def whiteningLocal(CC, yc, xc, nRange):
         # First channel in this list will always be the primary channel.
         ilocal = ilocal[:nRange]
 
-        wrot0 = cp.asnumpy(whiteningFromCovariance(CC[np.ix_(ilocal, ilocal)]))
+        wrot0 = cp.asnumpy(whiteningFromCovariance(CC[np.ix_(ilocal, ilocal)], epsilon=epsilon))
         # the first column of wrot0 is the whitening filter for the primary channel
         Wrot[ilocal, j] = wrot0[:, 0]
 
     return Wrot
 
 
-def get_whitening_matrix(raw_data=None, probe=None, params=None, nSkipCov=None):
+def get_data_covariance_matrix(raw_data, params, probe, nSkipCov=None, preprocessing_function=None):
+    """
+    Computes the data covariance matrix from a raw data file. Computes from a set of samples from the raw
+    data
+    :param raw_data:
+    :param params:
+    :param probe:
+    :param nSkipCov: for the orginal kilosort2, number of batches to skip between samples
+    :param preprocessing_function: if 'destriping' uses Neuropixel destriping function to apply to the data before
+    computing the covariance matrix
+    :return:
+    """
+    preprocessing_function = preprocessing_function or params.preprocessing_function
+    if preprocessing_function == 'destriping' and params.normalisation != "original":
+        # takes 25 samples of 500ms from 10 seconds to t -25s
+        good_channels = probe.good_channels
+        CCall = np.zeros((25, probe.Nchan, probe.Nchan))
+        t0s = np.linspace(10, raw_data.shape[0] / params.fs - 10, 25)
+        # on the second pass, apply destriping to the data
+        for icc, t0 in enumerate(tqdm(t0s, desc="Computing covariance matrix")):
+            s0 = slice(int(t0 * params.fs), int((t0 + 0.4) * params.fs))
+            raw = raw_data[s0][:, :probe.Nchan].T.astype(np.float32) * probe['sample2volt']
+            datr = destripe(raw, fs=params.fs, h=probe, channel_labels=probe.channels_labels) / probe['sample2volt']
+            assert not (np.any(np.isnan(datr)) or np.any(np.isinf(datr))), "destriping unexpectedly produced NaNs"
+            CCall[icc, :, :] = np.dot(datr, datr.T) / datr.shape[1]
+            # remove the bad channels from the covariance matrix, those get only divided by their rms
+            CCall[icc, :, :] = np.dot(datr, datr.T) / datr.shape[1]
+            CCall[icc, ~probe.good_channels, :] = 0
+            CCall[icc, :, ~probe.good_channels] = 0
+            # we stabilize the covariance matrix this is not trivial:
+            # first remove all cross terms belonging to the bad channels
+            median_std = np.median(np.diag(CCall[icc, :, :])[good_channels])
+            # channels flagged as noisy (label=1) and dead (label=2) are interpolated
+            replace_diag = np.interp(np.where(~good_channels)[0], np.where(good_channels)[0], np.diag(CCall[icc, :, :])[good_channels])
+            CCall[icc, ~good_channels, ~good_channels] = replace_diag
+            # the channels outside of the brain (label=3) are willingly excluded and stay with the high values above
+            CCall[icc, probe.channels_labels == 3, probe.channels_labels == 3] = median_std * 1e6
+        CC = cp.asarray(np.median(CCall, axis=0))
+    else:
+        nSkipCov = nSkipCov or params.nSkipCov
+        Nbatch = get_Nbatch(raw_data, params)
+        NT, NTbuff, ntbuff = (params.NT, params.NTbuff, params.ntbuff)
+        nbatches_cov = np.arange(0, Nbatch, nSkipCov).size
+        CCall = cp.zeros((nbatches_cov, probe.Nchan, probe.Nchan))
+        for icc, ibatch in enumerate(tqdm(range(0, Nbatch, nSkipCov), desc="Computing the whitening matrix")):
+            i = max(0, NT * ibatch - ntbuff)
+            # WARNING: we no longer use Fortran order, so raw_data is nsamples x NchanTOT
+            buff = raw_data[i:i + NTbuff]
+            assert buff.shape[0] > buff.shape[1]
+            assert buff.flags.c_contiguous
+            nsampcurr = buff.shape[0]
+            if nsampcurr < NTbuff:
+                buff = np.concatenate(
+                    (buff, np.tile(buff[nsampcurr - 1], (NTbuff - nsampcurr, 1))), axis=0)
+            buff_g = cp.asarray(buff, dtype=np.float32)
+            # apply filters and median subtraction
+            datr = gpufilter(buff_g, fs=params.fs, fshigh=params.fshigh, chanMap=probe.chanMap)
+            # remove buffers on either side of the data batch
+            datr = datr[ntbuff: NT + ntbuff]
+            assert datr.flags.c_contiguous
+            CCall[icc, :, :] = cp.dot(datr.T, datr) / datr.shape[0]
+        CC = cp.median(CCall, axis=0)
+    return CC
+
+
+def get_whitening_matrix(raw_data=None, probe=None, params=None, qc_path=None):
     """
     based on a subset of the data, compute a channel whitening matrix
     this requires temporal filtering first (gpufilter)
+    :param raw_data:
+    :param probe:
+    :param params:
+    :param kwargs: get_data_covariance_matrix kwargs
     """
-    Nbatch = get_Nbatch(raw_data, params)
-    ntbuff = params.ntbuff
-    NTbuff = params.NTbuff
-    whiteningRange = params.whiteningRange
-    scaleproc = params.scaleproc
-    NT = params.NT
-    fs = params.fs
-    fshigh = params.fshigh
-    if nSkipCov is None:
-        nSkipCov = params.nSkipCov
-
-    xc = probe.xc
-    yc = probe.yc
-    chanMap = probe.chanMap
     Nchan = probe.Nchan
-
-    # Nchan is obtained after the bad channels have been removed
-    CC = cp.zeros((Nchan, Nchan))
-
-    nbatches_cov = np.arange(0, Nbatch, nSkipCov).size
-    CCall = cp.zeros((nbatches_cov, Nchan, Nchan))
-
-    for icc, ibatch in enumerate(tqdm(range(0, Nbatch, nSkipCov), desc="Computing the whitening matrix")):
-        i = max(0, NT * ibatch - ntbuff)
-        # WARNING: we no longer use Fortran order, so raw_data is nsamples x NchanTOT
-        buff = raw_data[i:i + NTbuff]
-        assert buff.shape[0] > buff.shape[1]
-        assert buff.flags.c_contiguous
-
-        nsampcurr = buff.shape[0]
-        if nsampcurr < NTbuff:
-            buff = np.concatenate(
-                (buff, np.tile(buff[nsampcurr - 1], (NTbuff - nsampcurr, 1))), axis=0)
-
-        buff_g = cp.asarray(buff, dtype=np.float32)
-
-        # apply filters and median subtraction
-        datr = gpufilter(buff_g, fs=fs, fshigh=fshigh, chanMap=chanMap)
-
-        # remove buffers on either side of the data batch
-        datr = datr[ntbuff: NT + ntbuff]
-        assert datr.flags.c_contiguous
-        CCall[icc, :, :] = cp.dot(datr.T, datr) / datr.shape[0]
-    CC = cp.median(CCall, axis=0)
-
-    if params.do_whitening:
-        if whiteningRange < np.inf:
+    CC = get_data_covariance_matrix(raw_data, params, probe)
+    logger.info(f"Data normalisation using {params.normalisation} method")
+    if params.normalisation in ['whitening', 'original']:
+        epsilon = np.mean(np.diag(CC)[probe.good_channels]) * 1e-3 if params.normalisation == ['whitening'] else 1e-6
+        if params.whiteningRange < np.inf:
             #  if there are too many channels, a finite whiteningRange is more robust to noise
             # in the estimation of the covariance
-            whiteningRange = min(whiteningRange, Nchan)
+            whiteningRange = min(params.whiteningRange, Nchan)
             # this function performs the same matrix inversions as below, just on subsets of
             # channels around each channel
-            Wrot = whiteningLocal(CC, yc, xc, whiteningRange)
+            Wrot = whiteningLocal(CC, probe.yc, probe.xc, whiteningRange, epsilon=epsilon)
         else:
-            Wrot = whiteningFromCovariance(CC)
-    else:
-        # Do single channel z-scoring instead of whitening
+            Wrot = whiteningFromCovariance(CC, epsilon=epsilon)
+    elif params.normalisation == 'zscore':
+        # Do individual channel z-scoring instead of whitening
         Wrot = cp.diag(cp.diag(CC) ** (-0.5))
+    elif params.normalisation == 'global_zscore':
+        Wrot = cp.eye(CC.shape[0]) * np.median(cp.diag(CC) ** (-0.5))  # same value for all channels
+    if qc_path is not None:
+        pykilosort.qc.plot_whitening_matrix(Wrot.get(), good_channels=probe.good_channels, out_path=qc_path)
+        pykilosort.qc.plot_covariance_matrix(CC.get(), out_path=qc_path)
 
-    Wrot = Wrot * scaleproc
-    condition_number = np.linalg.cond(cp.asnumpy(Wrot))
+    Wrot = Wrot * params.scaleproc
+    condition_number = np.linalg.cond(cp.asnumpy(Wrot)[probe.good_channels, :][:, probe.good_channels])
     logger.info(f"Computed the whitening matrix cond = {condition_number}.")
     if condition_number > 50:
         logger.warning("high conditioning of the whitening matrix can result in noisy and poor results")
     return Wrot
 
-def get_good_channels(raw_data=None, probe=None, params=None):
+
+def get_good_channels(raw_data, probe, params, method='kilosort', **kwargs):
+    if method == 'raw_correlations':
+        return get_good_channels_raw_correlations(raw_data, probe, params, **kwargs)
+    else:
+        return get_good_channels_kilosort(raw_data, probe, params)
+
+
+def get_good_channels_raw_correlations(raw_data, params, probe, t0s=None, return_labels=False):
+    """
+    Detect bad channels using the method described in IBL whitepaper
+    :param raw_data:
+    :param params:
+    :param probe:
+    :param t0s:
+    :return:
+    """
+    if t0s is None:
+        t0s = np.linspace(10, raw_data.shape[0] / params.fs - 10, 25)
+    channel_labels = np.zeros((probe.Nchan, t0s.size))
+    for icc, t0 in enumerate(tqdm(t0s, desc="Auto-detection of noisy channels")):
+        s0 = slice(int(t0 * params.fs), int((t0 + 0.4) * params.fs))
+        raw = raw_data[s0][:, :probe.Nchan].T.astype(np.float32) * probe['sample2volt']
+        channel_labels[:, icc], _ = detect_bad_channels(raw, params.fs)
+    channel_labels = scipy.stats.mode(channel_labels, axis=1)[0].squeeze()
+    logger.info(f"Detected {np.sum(channel_labels == 1)} dead channels")
+    logger.info(f"Detected {np.sum(channel_labels == 2)} noise channels")
+    logger.info(f"Detected {np.sum(channel_labels == 3)} uppermost channels outside of the brain")
+    if return_labels:
+        return channel_labels == 0, channel_labels
+    else:
+        return channel_labels == 0
+
+
+def get_good_channels_kilosort(raw_data=None, probe=None, params=None):
     """
     of the channels indicated by the user as good (chanMap)
     further subset those that have a mean firing rate above a certain value